pyproject.toml
[project]
name = "riddle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.5",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]
